refactor(svelte2tsx): bind: ディレクティブを attributes/binding.rs へ切り出す
bind: の整形・フィルタ判定・要素変数名の生成一式を純粋移動。
上流の Binding.ts に対応させる。出力・ロジックの変更なし。
